refactor(in-app): move containers from dialog
