+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/dialog/alert/AlertDialog.kt
@@ -10,6 +10,7 @@ import android.view.ViewGroup
 import com.personalization.databinding.AlertDialogBinding
 import com.personalization.inAppNotification.view.component.container.InAppViewContainer
 import com.personalization.inAppNotification.view.component.dialog.BaseInAppDialog
+import com.personalization.inAppNotification.view.component.dialog.alert.container.AlertDialogViewContainer
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_COLOR_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_TEXT_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_POSITIVE_COLOR_KEY
@@ -24,18 +25,6 @@ class AlertDialog : BaseInAppDialog() {
 
     private var _binding: AlertDialogBinding? = null
     private val binding get() = _binding!!
-
-    class AlertDialogViewContainer(binding: AlertDialogBinding) : InAppViewContainer {
-        override val backgroundImageView = binding.backgroundImageView
-        override val imageContainer = binding.imageContainer
-        override val titleTextView = binding.textContainer.title
-        override val messageTextView = binding.textContainer.message
-        override val buttonAcceptContainer = binding.buttonContainer.buttonAcceptContainer
-        override val buttonAccept = binding.buttonContainer.buttonAccept
-        override val buttonDeclineContainer = binding.buttonContainer.buttonDeclineContainer
-        override val buttonDecline = binding.buttonContainer.buttonDecline
-        override val closeButton = binding.closeButton
-    }
 
     override val container: InAppViewContainer by lazy {
         AlertDialogViewContainer(binding)
