+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/dialog/bottomSheet/BottomSheetDialog.kt
@@ -12,6 +12,7 @@ import android.view.ViewGroup
 import com.personalization.databinding.BottomSheetDialogBinding
 import com.personalization.inAppNotification.view.component.container.InAppViewContainer
 import com.personalization.inAppNotification.view.component.dialog.BaseInAppDialog
+import com.personalization.inAppNotification.view.component.dialog.bottomSheet.container.BottomSheetDialogViewContainer
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_COLOR_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_TEXT_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_POSITIVE_COLOR_KEY
@@ -27,20 +28,8 @@ class BottomSheetDialog : BaseInAppDialog() {
     private var _binding: BottomSheetDialogBinding? = null
     private val binding get() = _binding!!
 
-    class BottomSheetViewContainer(binding: BottomSheetDialogBinding) : InAppViewContainer {
-        override val backgroundImageView = binding.backgroundImageView
-        override val imageContainer = binding.imageContainer
-        override val titleTextView = binding.textContainer.title
-        override val messageTextView = binding.textContainer.message
-        override val buttonAcceptContainer = binding.buttonContainer.buttonAcceptContainer
-        override val buttonAccept = binding.buttonContainer.buttonAccept
-        override val buttonDeclineContainer = binding.buttonContainer.buttonDeclineContainer
-        override val buttonDecline = binding.buttonContainer.buttonDecline
-        override val closeButton = binding.closeButton
-    }
-
     override val container: InAppViewContainer by lazy {
-        BottomSheetViewContainer(binding)
+        BottomSheetDialogViewContainer(binding)
     }
 
     override fun onCreateView(
