+++ personalization-sdk/src/main/kotlin/com/personalization/inAppNotification/view/component/dialog/topSheet/TopSheetDialog.kt
@@ -12,6 +12,7 @@ import android.view.ViewGroup
 import com.personalization.databinding.TopSheetDialogBinding
 import com.personalization.inAppNotification.view.component.container.InAppViewContainer
 import com.personalization.inAppNotification.view.component.dialog.BaseInAppDialog
+import com.personalization.inAppNotification.view.component.dialog.topSheet.container.TopSheetDialogViewContainer
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_COLOR_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_NEGATIVE_TEXT_KEY
 import com.personalization.inAppNotification.view.component.utils.InAppConsts.BUTTON_POSITIVE_COLOR_KEY
@@ -27,21 +28,8 @@ class TopSheetDialog : BaseInAppDialog() {
     private var _binding: TopSheetDialogBinding? = null
     private val binding get() = _binding!!
 
-    class TopSheetViewContainer(binding: TopSheetDialogBinding) : InAppViewContainer {
-        override val backgroundImageView = binding.backgroundImageView
-        override val imageContainer = binding.imageContainer
-        override val titleTextView = binding.textContainer.title
-        override val messageTextView = binding.textContainer.message
-        override val buttonAcceptContainer = binding.buttonContainer.buttonAcceptContainer
-        override val buttonAccept = binding.buttonContainer.buttonAccept
-        override val buttonDeclineContainer = binding.buttonContainer.buttonDeclineContainer
-        override val buttonDecline = binding.buttonContainer.buttonDecline
-        override val closeButton = binding.closeButton
-    }
-
-
     override val container: InAppViewContainer by lazy {
-        TopSheetViewContainer(binding)
+        TopSheetDialogViewContainer(binding)
     }
 
     override fun onCreateView(
